requirements.txt
libusb1==1.6.6
